requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
bitarray==3.7.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ckzg==2.1.4
click==8.2.1
cryptography==46.0.1
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.4
hexbytes==1.3.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.4
parsimonious==0.10.0
propcache==0.3.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pyunormalize==16.0.0
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
SQLAlchemy==2.0.41
toolz==1.0.0
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
web3==7.13.0

websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
pytest-asyncio==0.23.2
